Updated profile.js
Made the necessary changes to integrate my 'Adjust Basal Inverse to Autosens ISF' middleware directly into the apps code and added a toggle to Preferences in the 'OPENAPS MAIN SETTINGS' section to enable or disable it, disabled my default.

Made the necessary changes to integrate an 'Use Autosens ISF to Calculate the autoISF Sens' toggle into autoISF settings just below 'Enable Autosens'. When this is toggled on, it automatically toggles Enable Autosens to true if it isn't already and uses autosens_data.newisf instead of profile.sens to determine the autoISF final sens.